requirements.txt
icalendar==5.0.12
recurring-ical-events==2.1.2
pytz==2024.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
Jinja2==3.1.4
requests==2.32.3
